AS Levels redone, removed outdated Seed Generator files
