+++ Scripts/AS/ASGR_hooks.lua
@@ -1,0 +1,13 @@
+function ASGR_hooks(Ob)
+
+    local onBeginLevel_original = Ob.onBeginLevel
+
+    function Ob:onBeginLevel()
+        %onBeginLevel_original(self)
+        GamePrint("Open the Gate!")
+		--edit to make gate always open
+		if (self.gate) then
+			self:sendMessage(self.gate, 'Open')
+		end
+    end
+end
