--- SeedGenerator/SettingsPresets/allgoals.ini
@@ -1,50 +1,0 @@
-[SeedSettings]
-startcobwebduster = False
-
-startlevitation = True
-
-startbutton = False
-
-randomizecobwebduster = True
-
-everylocationpossible = True
-
-instantdeath = False
-
-[VictoryConditions]
-beatoleander = True
-
-rank101 = True
-
-brainhunt = True
-
-scavengerhunt = True
-
-beatalllevels = True
-
-[QualityOfLife]
-fasterLO = True
-
-easymillarace = True
-
-earlyelevator = True
-
-mentalmagnet = True
-
-[AdditionalFiles]
-spoilerlog = True
-
-createhints = True
-
-detailedhints = True
-
-#Minimum Value 0, Maximum Value 7
-majorhints = 7
-
-#Minimum Value 0, Maximum Value 8
-powerhints = 8
-
-#Minimum Value 0, Maximum Value 11
-minorhints = 11
-
-
