--- SeedGenerator/SettingsPresets/beginner.ini
@@ -1,50 +1,0 @@
-[SeedSettings]
-startcobwebduster = True
-
-startlevitation = True
-
-startbutton = True
-
-randomizecobwebduster = False
-
-everylocationpossible = True
-
-instantdeath = False
-
-[VictoryConditions]
-beatoleander = True
-
-rank101 = False
-
-brainhunt = False
-
-scavengerhunt = False
-
-beatalllevels = False
-
-[QualityOfLife]
-fasterLO = True
-
-easymillarace = True
-
-earlyelevator = True
-
-mentalmagnet = True
-
-[AdditionalFiles]
-spoilerlog = True
-
-createhints = True
-
-detailedhints = True
-
-#Minimum Value 0, Maximum Value 7
-majorhints = 3
-
-#Minimum Value 0, Maximum Value 8
-powerhints = 3
-
-#Minimum Value 0, Maximum Value 11
-minorhints = 4
-
-
