--- SeedGenerator/SettingsPresets/rank101.ini
@@ -1,50 +1,0 @@
-[SeedSettings]
-startcobwebduster = False
-
-startlevitation = True
-
-startbutton = False
-
-randomizecobwebduster = True
-
-everylocationpossible = True
-
-instantdeath = False
-
-[VictoryConditions]
-beatoleander = False
-
-rank101 = True
-
-brainhunt = False
-
-scavengerhunt = False
-
-beatalllevels = False
-
-[QualityOfLife]
-fasterLO = True
-
-easymillarace = True
-
-earlyelevator = True
-
-mentalmagnet = True
-
-[AdditionalFiles]
-spoilerlog = True
-
-createhints = False
-
-detailedhints = False
-
-#Minimum Value 0, Maximum Value 7
-majorhints = 0
-
-#Minimum Value 0, Maximum Value 8
-powerhints = 0
-
-#Minimum Value 0, Maximum Value 11
-minorhints = 0
-
-
