--- SeedGenerator/SpoilerLogFiles/item_spoiler.py
@@ -1,215 +1,0 @@
-item_spoiler_list = [
-#Current Props from AS, 6 total
-'Lungfish Call', 'Glorias Trophy', 'Straight Jacket', 'Loboto Painting', 'Cake', 'Lilis Bracelet',
-
-#Current Props from MM, 6 total
-'Prop Sign', 'Prop Flowers', 'Prop Plunger', 'Prop HedgeTrimmers', 'Prop RollingPin', 'Prop WaterCan',
-
-#Current Props from TH, 3 total
-'Candle1', 'Candle2', 'Megaphone',
-
-#Props from WW, 3 total
-'Freds Note', #Fred's Note
-'Priceless Coin', #Priceless Coin
-'Rifle', #Rifle
-
-#19 Psi Powers, Some Different Names to match Class
-'Marksmanship1',
-'Marksmanship2',
-'Marksmanship3',
-'Pyrokinesis1',
-'Pyrokinesis2',
-'Confusion1',
-'Confusion2',
-'Levitation1',
-'Levitation2',
-'Levitation3',
-'Telekinesis1',
-'Telekinesis2',
-'Invisibility1',
-'Invisibility2',
-'Clairvoyance1',
-'Clairvoyance2',
-'Shield1',
-'Shield2',
-'Shield3',
-
-#6 Max Ammo Up
-'AmmoUp1',
-'AmmoUp2',
-'AmmoUp3',
-'AmmoUp4',
-'AmmoUp5',
-'AmmoUp6',
-
-#6 Max Lives Up
-'MaxLivesUp1',
-'MaxLivesUp2',
-'MaxLivesUp3',
-'MaxLivesUp4',
-'MaxLivesUp5',
-'MaxLivesUp6',
-
-#4 Confusion Ammo Up
-'ConfusionUp1',
-'ConfusionUp2',
-'ConfusionUp3',
-'ConfusionUp4',
-
-#10 Challenge ChallengeMarkers
-'ChallengeMarker1',
-'ChallengeMarker2',
-'ChallengeMarker3',
-'ChallengeMarker4',
-'ChallengeMarker5',
-'ChallengeMarker6',
-'ChallengeMarker7',
-'ChallengeMarker8',
-'ChallengeMarker9',
-'ChallengeMarker10',
-           
-#19 Brain Jars, Unique Names
-'BrainJarElton',
-'BrainJarBobby',
-'BrainJarDogen',
-'BrainJarBenny',
-'BrainJarElka',
-'BrainJarKitty',
-'BrainJarChloe',
-'BrainJarFranke',
-'BrainJarJT',
-'BrainJarQuentin',
-'BrainJarVernon',
-'BrainJarMilka',
-'BrainJarCrystal',
-'BrainJarClem',
-'BrainJarNils',
-'BrainJarMaloof',
-'BrainJarMikhail',
-'BrainJarPhoebe',
-'BrainJarChops',
-
-#16 Scavenger Hunt Items, Unique Names, in Order
-'Gold Doubloon',
-'Eagle Claw',
-'Divers Helmet',
-'Psy Comic',
-'Wood Pipe',
-'Turkey Sandwich',
-'Voodoo Doll',
-'Miner Skull',
-'Pirate Scope',
-'Golden Acorn',
-'Glass Eye',
-'Egg',
-'Fertility Idol',
-'Dinosaur Bone',
-'Fossil',
-'Gold Watch',
-
-#50 Emotional Baggage Tags, 10 of each Type
-'SuitcaseTag1',
-'SuitcaseTag2',
-'SuitcaseTag3',
-'SuitcaseTag4',
-'SuitcaseTag5',
-'SuitcaseTag6',
-'SuitcaseTag7',
-'SuitcaseTag8',
-'SuitcaseTag9',
-'SuitcaseTag10',
-
-'PurseTag1',
-'PurseTag2',
-'PurseTag3',
-'PurseTag4',
-'PurseTag5',
-'PurseTag6',
-'PurseTag7',
-'PurseTag8',
-'PurseTag9',
-'PurseTag10',
-
-'HatboxTag1',
-'HatboxTag2',
-'HatboxTag3',
-'HatboxTag4',
-'HatboxTag5',
-'HatboxTag6',
-'HatboxTag7',
-'HatboxTag8',
-'HatboxTag9',
-'HatboxTag10',
-
-'SteamerTag1',
-'SteamerTag2',
-'SteamerTag3',
-'SteamerTag4',
-'SteamerTag5',
-'SteamerTag6',
-'SteamerTag7',
-'SteamerTag8',
-'SteamerTag9',
-'SteamerTag10',
-
-'DuffleTag1',
-'DuffleTag2',
-'DuffleTag3',
-'DuffleTag4',
-'DuffleTag5',
-'DuffleTag6',
-'DuffleTag7',
-'DuffleTag8',
-'DuffleTag9',
-'DuffleTag10',
-
-#50 Emotional Baggage, 10 of each Type
-'Suitcase1', 'Suitcase2', 'Suitcase3', 'Suitcase4', 'Suitcase5', 'Suitcase6', 'Suitcase7', 'Suitcase8', 'Suitcase9', 'Suitcase10',
-
-'Purse1', 'Purse2', 'Purse3', 'Purse4', 'Purse5', 'Purse6', 'Purse7', 'Purse8', 'Purse9', 'Purse10',
-
-'Hatbox1', 'Hatbox2', 'Hatbox3', 'Hatbox4', 'Hatbox5', 'Hatbox6', 'Hatbox7', 'Hatbox8','Hatbox9','Hatbox10',
-
-'Steamertrunk1', 'Steamertrunk2', 'Steamertrunk3', 'Steamertrunk4', 'Steamertrunk5', 'Steamertrunk6', 'Steamertrunk7', 'Steamertrunk8', 'Steamertrunk9', 'Steamertrunk10',
-
-'Dufflebag1', 'Dufflebag2', 'Dufflebag3', 'Dufflebag4', 'Dufflebag5', 'Dufflebag6', 'Dufflebag7', 'Dufflebag8', 'Dufflebag9', 'Dufflebag10',
-
-#19 Vaults
-'Vault1', 'Vault2', 'Vault3', 'Vault4', 'Vault5', 'Vault6', 'Vault7', 'Vault8', 'Vault9', 'Vault10',
-'Vault11', 'Vault12', 'Vault13', 'Vault14', 'Vault15', 'Vault16', 'Vault17','Vault18', 'Vault19',
-
-#30 Rando Arrowhead Bundles Small
-'AHSmall1', 'AHSmall2', 'AHSmall3', 'AHSmall4', 'AHSmall5', 'AHSmall6', 'AHSmall7', 'AHSmall8', 'AHSmall9', 'AHSmall10',
-'AHSmall11', 'AHSmall12', 'AHSmall13', 'AHSmall14', 'AHSmall15', 'AHSmall16', 'AHSmall17', 'AHSmall18', 'AHSmall19', 'AHSmall20',
-'AHSmall21', 'AHSmall22', 'AHSmall23', 'AHSmall24', 'AHSmall25', 'AHSmall26', 'AHSmall27', 'AHSmall28', 'AHSmall29', 'AHSmall30', 
-
-#5 Rando Arrowhead Bundles Large
-'AHLarge1', 'AHLarge2', 'AHLarge3', 'AHLarge4', 'AHLarge5',
-
-#Oarsman's Badge
-'Oarsmans Badge',
-
-#Sasha's Button
-'Button',
-
-#Crow Feather
-'Feather',
-
-#Cobweb Duster
-'Rando Cobweb Duster',
-
-#111 Psicards, filler item, increase if adding more positions
-'Card1', 'Card2', 'Card3', 'Card4', 'Card5', 'Card6', 'Card7', 'Card8', 'Card9', 'Card10',
-'Card11', 'Card12', 'Card13', 'Card14', 'Card15', 'Card16', 'Card17', 'Card18', 'Card19', 'Card20',
-'Card21', 'Card22', 'Card23', 'Card24', 'Card25', 'Card26', 'Card27', 'Card28', 'Card29', 'Card30', 
-'Card31', 'Card32', 'Card33', 'Card34', 'Card35', 'Card36', 'Card37', 'Card38', 'Card39', 'Card40',
-'Card41', 'Card42', 'Card43', 'Card44', 'Card45', 'Card46', 'Card47', 'Card48', 'Card49', 'Card50',
-'Card51', 'Card52', 'Card53', 'Card54', 'Card55', 'Card56', 'Card57', 'Card58', 'Card59', 'Card60',
-'Card61', 'Card62', 'Card63', 'Card64', 'Card65', 'Card66', 'Card67', 'Card68', 'Card69', 'Card70',
-'Card71', 'Card72', 'Card73', 'Card74', 'Card75', 'Card76', 'Card77', 'Card78', 'Card79', 'Card80', 
-'Card81', 'Card82', 'Card83', 'Card84', 'Card85', 'Card86', 'Card87', 'Card88', 'Card89', 'Card90', 
-'Card91', 'Card92', 'Card93', 'Card94', 'Card95', 'Card96', 'Card97', 'Card98', 'Card99', 'Card100',
-'Card101', 'Card102', 'Card103', 'Card104', 'Card105', 'Card106', 'Card107', 'Card108', 'Card109', 'Card110',
-'Card111',
-
-]
