--- SeedGenerator/SpoilerLogFiles/location_spoiler.py
@@ -1,431 +1,0 @@
-location_spoiler_list = [
-
-# WHISPERING ROCK SUMMER CAMP
-# Sasha's Underground Lab [CASA] 
-"[CASA] Behind Furniture (Card)",
-"[CASA] Staircase Ledges (Card)",
-"[CASA] Upper Ledge (Fossil)",
-# GPC and Wilderness [CAGP] 
-"[CAGP] Top of GPC (Card)",
-"[CAGP] Under GPC (Card)",
-"[CAGP] Mountain Lion Log Bridge (Card)",
-"[CAGP] Above Entrance to Lake (Card)",
-"[CAGP] Rock Wall Behind Tree (Card)",
-"[CAGP] Rock Wall Top (Pirate Scope)",
-"[CAGP] Tree Near Fence (Card)",
-"[CAGP] Tree Near Geyser (Card)",
-"[CAGP] Fence Behind GPC (Card)",
-"[CAGP] Near the Bear (Card)",
-"[CAGP] Rocky Platforms Behind GPC Right (Card)",
-"[CAGP] Rocky Platforms Behind GPC Left (Card)",
-"[CAGP] Top of Log Flume (Card)",
-"[CAGP] Ride the Log Flume (Card)",
-"[CAGP] Bottom of Log Flume (Card)",
-"[CAGP] Big Rock Near Ford (Card)",
-"[CAGP] Ruined Cabin (Challenge Marker)",
-"[CAGP] Branch Swinging Course Start (Card)",
-"[CAGP] Branch Swinging Course Mid (Card)",
-"[CAGP] Branch Swinging Course End (Challenge Marker)",
-"[CAGP] Branch Above Squirrel (Card)",
-"[CAGP] Creek Grate (Glass Eye)",
-"[CAGP] Squirrel's Acorn (Golden Acorn)",
-"[CAGP] Geyser (Miner's Skull)",
-# Main Campgrounds [CAMA] 
-"[CAMA] Fence Near Kids' Cabins (Card)",
-"[CAMA] Under Lodge Front Steps (Card)",
-"[CAMA] Behind Tree Near Lodge (Card)",
-"[CAMA] Under the Lodge (Gold Doubloon)",
-"[CAMA] Loudspeaker 1 Platform  (Card)",
-"[CAMA] Under Lodge Metal Roof (Card)",
-"[CAMA] Loudspeaker Tightrope Walk (Card)",
-"[CAMA] Loudspeaker 2 Platform (Card)",
-"[CAMA] Lodge Roof (Challenge Marker)",
-"[CAMA] Metal Roof Outcropping (Card)",
-"[CAMA] Loudspeaker Above Stump (Card)",
-"[CAMA] Tree Platform Left (Card)",
-"[CAMA] Tree Platform Right (Eagle Claw)",
-"[CAMA] Rock Wall Top (Card)",
-"[CAMA] Parking Lot Arch (Card)",
-"[CAMA] Parking Lot Small Log (Card)",
-"[CAMA] Oleander's Car (Card)",
-"[CAMA] Parking Lot Basketball Hoop (Card)",
-"[CAMA] Parking Lot History Board (Card)",
-"[CAMA] Parking Lot Outhouse (Card)",
-"[CAMA] Rock Near Bench (Card)",
-# Kids' Cabins [CAKC] 
-"[CAKC] Grinding on the Roots (Card)",
-"[CAKC] Under Stairs (Card)",
-"[CAKC] Top of Loudspeaker (Card)",
-"[CAKC] Cabin Roof 1 (Card)",
-"[CAKC] Trampoline Above Outhouse (Card)",
-"[CAKC] Trampoline Platform (Challenge Marker)",
-"[CAKC] Cabins Outhouse (Card)",
-"[CAKC] Behind Cabin (Card)",
-"[CAKC] Roof of Cabin 2 (Card)",
-"[CAKC] Cave Entrance (Card)",
-"[CAKC] Deep Cave Path (Card)",
-"[CAKC] Deep Cave Ladder (Card)",
-"[CAKC] High-Up Tightrope (Card)",
-"[CAKC] Cave Refrigerator (Turkey Sandwich)",
-# Reception Area [CARE] 
-"[CARE] Graveyard Bear (Card)",
-"[CARE] Near Beehive (Card)",
-"[CARE] Mineshaft Trailer Entrance (Card)",
-"[CARE] Tightrope Start (Card)",
-"[CARE] Tightrope End (Card)",
-"[CARE] Rocks Near Trailer (Card)",
-"[CARE] Fireplace Tree Lower (Card)",
-"[CARE] Fireplace Tree Rock (Card)",
-"[CARE] Swamp Skinny Poles (Card)",
-"[CARE] Big Log Platform (Card)",
-"[CARE] Above Waterfall Left (Card)",
-"[CARE] Above Waterfall Right (Card)",
-"[CARE] Behind the Waterfall (Card)",
-"[CARE] Weird Tree Left (Cherry Wood Pipe)",
-"[CARE] Weird Tree Right (Card)",
-"[CARE] Log Hill Top (Card)",
-"[CARE] Log Hill Behind (Card)",
-"[CARE] Mineshaft Grind Rail (Card)",
-"[CARE] Mineshaft Upper Entrance (Card)",
-"[CARE] Mineshaft Above Upper Entrance (Card)",
-"[CARE] Inside Mineshaft (Card)",
-"[CARE] Mineshaft Bear (Card)",
-"[CARE] Swamp Bird's Nest (Condor Egg)",
-"[CARE] Collapsed Cave (Challenge Marker)",
-"[CARE] Fireplace Tree Top (Dinosaur Bone)",
-"[CARE] Hornet Nest (Fertility Idol)",
-# Lake and Boathouse [CABH] 
-"[CABH] Under the First Bridge (Card)",
-"[CABH] Behind Stump (Card)",
-"[CABH] Left of Entrance Rock Wall (Card)",
-"[CABH] Poles on Lake (Card)",
-"[CABH] Bathysphere Roof (Card)",
-"[CABH] Bathysphere Dock (Card)",
-"[CABH] Metal Roof Above Ford (Card)",
-"[CABH] Above Ford Ropes (Card)",
-"[CABH] Above Ford Cabin Platform (Card)",
-"[CABH] Outside Cougar Cave (Card)",
-"[CABH] Inside Cougar Cave (Diver's Helmet)",
-"[CABH] Bulletin Board Bushes (Card)",
-"[CABH] Pink Trees Platform Left (Card)",
-"[CABH] Pink Trees Platform Right (Card)",
-"[CABH] Rock Wall Upper (Card)",
-"[CABH] Lake Shore (Card)",
-"[CABH] Tiny Island (Card)",
-"[CABH] Top of Big Rock (Challenge Marker)",
-"[CABH] Rock Wall Gap (Psychonauts Comic # 1)",
-# Main Lodge [CALI] 
-"[CALI] Main Lodge Rafters (Voodoo Doll)",
-# Ford's Sanctuary [CAJA] 
-"[CAJA] Top of Sanctuary (Card)",
-"[CAJA] Bottom of Sanctuary (Card)",
-# Rando Only Rank Checks [RANK] 
-"[RANK] PSI Rank 05",
-"[RANK] PSI Rank 10",
-"[RANK] PSI Rank 15",
-"[RANK] PSI Rank 20",
-"[RANK] PSI Rank 25",
-"[RANK] PSI Rank 30",
-"[RANK] PSI Rank 35",
-"[RANK] PSI Rank 40",
-"[RANK] PSI Rank 45",
-"[RANK] PSI Rank 50",
-"[RANK] PSI Rank 55",
-"[RANK] PSI Rank 60",
-"[RANK] PSI Rank 65",
-"[RANK] PSI Rank 70",
-"[RANK] PSI Rank 75",
-"[RANK] PSI Rank 80",
-"[RANK] PSI Rank 85",
-"[RANK] PSI Rank 90",
-"[RANK] PSI Rank 95",
-"[RANK] PSI Rank 101",
-
-# THORNEY TOWERS
-# Asylum Grounds [ASGR] 
-"[ASGR] Rock Wall Bottom (Lili's Bracelet)",
-"[ASGR] Rock Wall Ladder (Card)",
-"[ASGR] Outside Front Gate (Challenge Marker)",
-"[ASGR] Pillar Above Gate (Card)",
-"[ASGR] Fountain Top (Gold Watch)",
-"[ASGR] Hedge Alcove (JT's Brain)",
-"[ASGR] Asylum Doors Right (Franke's Brain)",
-"[ASGR] Asylum Doors Left (Card)",
-"[ASGR] Corner Near Fence (Card)",
-"[ASGR] Ledge Before Gloria (Kitty's Brain)",
-# Asylum Lower Level [ASCO] 
-"[ASCO] Above Elevator (Card)",
-"[ASCO] Crow's Basket (Milka's Brain)",
-"[ASCO] Ledge Above Fred Left (Chops' Brain)",
-"[ASCO] Ledge Above Fred Right (Card)",
-"[ASCO] Ledge Opposite Elevator (Card)",
-"[ASCO] Edgar's Room (Card)",
-"[ASCO] Behind Elevator (Maloof's Brain)",
-"[ASCO] Junk Corner (Chloe's Brain)",
-"[ASCO] Above Edgar (Challenge Marker)",
-# Asylum Upper Level [ASUP] 
-"[ASUP] Behind Mattress Wall (Card)",
-"[ASUP] Checkered Bathroom (Benny's Brain)",
-"[ASUP] Room Near Checkered Bathroom (Card)",
-"[ASUP] Elevator Shaft (Card)",
-"[ASUP] Room Left of Pipe Slide (Bobby's Brain)",
-"[ASUP] Floating in Hole (Card)",
-"[ASUP] Next to Hole (Card)",
-"[ASUP] Crumbling Outer Wall Planks (Elka's Brain)",
-"[ASUP] Crumbling Outer Wall Pillar (Card)",
-"[ASUP] Crumbling Outer Wall Below Platform (Card)",
-"[ASUP] Crumbling Outer Wall Platform (Card)",
-"[ASUP] Room Above Tilted Stairs (Card)",
-"[ASUP] Acid Room Floor (Card)",
-"[ASUP] Acid Room Table (Card)",
-"[ASUP] Acid Room Window (Quentin's Brain)",
-"[ASUP] Acid Room Overhang (Card)",
-"[ASUP] Small Windows Ledge (Dogen's Brain)",
-"[ASUP] Round Wood Platform (Card)",
-"[ASUP] Grate Climb Bottom (Phoebe's Brain)",
-"[ASUP] Grate Climb Mid (Card)",
-"[ASUP] Sink Platform Left (Card)",
-"[ASUP] Sink Platform Right (Crystal's Brain)",
-"[ASUP] Pipes Below Chair Door (Challenge Marker)",
-"[ASUP] Room Opposite Chair Door (Clem's Brain)",
-"[ASUP] Pipe Slide Near Chair Door (Card)",
-"[ASUP] Rafters Above Chair Door (Nil's Brain)",
-# Dr. Loboto's Laboratory [ASLB] 
-"[ASLB] Lab Caged Crow Left (Card)",
-"[ASLB] Lab Caged Crow Right (Vernon's Brain)",
-"[ASLB] Next to Pokeylope (Card)",
-"[ASLB] Lab Top Railing Left 1 (Card)",
-"[ASLB] Lab Top Railing Left 2 (Jarred Brain)",
-"[ASLB] Lab Top Elevator (Jarred Brain)",
-"[ASLB] Lab Top Railing Right (Card)",
-"[ASLB] Tea Room (Birthday Cake)",
-
-# BASIC BRAINING
-# Obstacle Course 1 [BBA1] 
-"[BBA1] Jumping Tutorial 1 (Steamer Trunk)",
-"[BBA1] Jumping Tutorial 2 (Steamer Trunk Tag)",
-"[BBA1] Pole-Climbing Tutorial Floor (Suitcase Tag)",
-"[BBA1] Below the Triple Trampolines (Suitcase)",
-# Obstacle Course 2 [BBA2] 
-"[BBA2] Giant Soldier Cut-Out (Vault)",
-"[BBA2] Dodging Bullets 1 (Hatbox Tag)",
-"[BBA2] Dodging Bullets 2 (Hatbox)",
-"[BBA2] Machine Gun Turret (Dufflebag Tag)",
-"[BBA2] Pole-Swinging Tutorial (Dufflebag)",
-"[BBA2] Trapeze Cobweb (Vault)",
-"[BBA2] Trapeze Platform (Purse Tag)",
-"[BBA2] Inside Plane Wreckage (Purse)",
-# Obstacle Course Finale [BBLT] 
-"[BBLT] End of Obstacle Course Left",
-"[BBLT] End of Obstacle Course Right",
-"[BBLT] Basic Braining Complete",
-
-# SASHA'S SHOOTING GALLERY
-# Cube Face 1 [SACU] 
-"[SACU] On the Bed (Memory Vault)",
-"[SACU] On the Pillow (Dufflebag Tag)",
-"[SACU] Building Blocks Left (Hatbox Tag)",
-"[SACU] Building Blocks Below (Dufflebag)",
-"[SACU] Building Blocks Right (One-Up)",
-"[SACU] Top of Bed Frame (Hatbox)",
-# Cube Face 2 [SACU] 
-"[SACU] Round Platforms Bottom (Purse Tag)",
-"[SACU] Round Platforms Near Valve (Purse)",
-"[SACU] Round Platforms Far from Valve (One-Up)",
-# Cube Face 3 [SACU] 
-"[SACU] Side of Cube Face 3 (Vault)",
-"[SACU] Bottom of Shoebox Ladder (Steamer Trunk Tag)",
-"[SACU] Shoebox Pedestal (Steamer Trunk)",
-"[SACU] Shoebox Tower Top (One-Up)",
-# Cube Face 4 [SACU] 
-"[SACU] Flame Tower Steps (Suitcase Tag)",
-"[SACU] Flame Tower Top 1 (Suitcase)",
-"[SACU] Flame Tower Top 2 (One-Up)",
-"[SACU] Sasha's Shooting Gallery Complete",
-
-# MILLA'S DANCE PARTY
-# The Lounge [MIFL] 
-"[MIFL] Intro Rings Tutorial (Dufflebag Tag)",
-"[MIFL] Dancing Camper Platform 1 (Dufflebag)",
-"[MIFL] Demon Room (Memory Vault)",
-"[MIFL] Windy Ladder Bottom (Steamer Trunk Tag)",
-"[MIFL] Pinball Plunger (Hatbox Tag)",
-"[MIFL] Plunger Party Ledge (Steamer Trunk)",
-"[MIFL] Grindrail Rings (Hatbox)",
-"[MIFL] Censor Hallway (Purse Tag)",
-"[MIFL] Pink Bowl Bottom (Vault)",
-"[MIFL] Pink Bowl Small Platform (Purse)",
-# The Party [MILL] 
-"[MILL] Bubbly Fan Bottom (Suitcase Tag)",
-"[MILL] Bubbly Fan Platform (Ammo Upgrade)",
-"[MILL] Bubbly Fan Top (Suitcase)",
-"[MILL] Milla's Party Room",
-"[MILL] Milla's Dance Party Complete",
-
-# BRAIN TUMBLER EXPERIMENT
-# The Nightmare Woods [NIMP] 
-"[NIMP] Outside Caravan (Vault)",
-"[NIMP] Behind the Egg (Purse Tag)",
-"[NIMP] Shadow Monster Path (Steamer Trunk Tag)",
-"[NIMP] Shadow Monster Blue Mushrooms (Purse)",
-"[NIMP] Ledge Behind Shadow Monster (Steamer Trunk)",
-"[NIMP] Below the Steep Ledge (Dufflebag Tag)",
-"[NIMP] Forest Path Blue Mushrooms (Suitcase Tag)",
-"[NIMP] Forest Blue Ledge (Suitcase)",
-"[NIMP] Forest High Platform (Dufflebag)",
-"[NIMP] Forest Path Thorns (Hatbox Tag)",
-"[NIMP] Behind Thorn Tower Left (Hatbox)",
-"[NIMP] Behind Thorn Tower Mid (Max Lives)",
-"[NIMP] Behind Thorn Tower Right (Vault)",
-# The Brain Tank [NIMP] 
-"[NIMP] Brain Tumbler Experiment Complete",
-
-# LUNGFISHOPOLIS
-# Lungfishopolis City [LOMA] 
-"[LOMA] Skyscraper Start (One Up)",
-"[LOMA] Corner Near Jail (Dufflebag)",
-"[LOMA] Skyscraper before Dam (Purse Tag)",
-"[LOMA] Behind Lasers Left 1 (Dufflebag Tag)",
-"[LOMA] Behind Lasers Left 2 (Vault)",
-"[LOMA] Behind Lasers Right (Purse)",
-"[LOMA] Blimp Hop (Hatbox Tag)",
-"[LOMA] End of Dam (Hatbox)",
-"[LOMA] End of Dam Platform (One Up)",
-"[LOMA] Skyscraper after Dam (Ammo Up)",
-"[LOMA] Near Battleships (Steamertrunk Tag)",
-"[LOMA] On the Bridge (One Up)",
-"[LOMA] Ground after Bridge (Vault)",
-"[LOMA] Skyscraper after Bridge (Steamertrunk)",
-"[LOMA] Tunnel (Suitcase Tag)",
-"[LOMA] Final Skyscrapers Left (Max Lives)",
-"[LOMA] Final Skyscrapers Right (Suitcase)",
-# Kochamara Arena [LOCB] 
-"[LOCB] Kochamara Intro Left",
-"[LOCB] Kochamara Intro Right",
-"[LOCB] Lungfishopolis Complete",
-
-# THE MILKMAN CONSPIRACY
-# The Neighborhood 1 [MMI1] 
-"[MMI1] Boyd's Fridge (Clairvoyance)",
-"[MMI1] First House (Dufflebag Tag)",
-"[MMI1] Second House (Rolling Pin)",
-"[MMI1] Car Trunk 1 (Stop Sign)",
-"[MMI1] Roof after Road Crew  (Purse Tag)",
-"[MMI1] Car Trunk 2 (Hedge Trimmers)",
-"[MMI1] Car House Backyard (Steamertrunk Tag)",
-"[MMI1] Inside Webbed Garage (Hatbox)",
-"[MMI1] Graveyard Patio (Vault)",
-"[MMI1] Graveyard Behind Tree (One Up)",
-"[MMI1] Behind Graveyard (Dufflebag)",
-"[MMI1] Hedge Maze (Flowers)",
-"[MMI1] Car Trunk 3 (Watering Can)",
-"[MMI1] Post Office Roof (One Up)",
-"[MMI1] Post Office Lobby (Suitcase)",
-"[MMI1] Post Office Basement (Plunger)",
-"[MMI1] Landscapers House Backyard (Suitcase Tag)",
-"[MMI1] Landscapers House Table (Purse)",
-"[MMI1] Landscapers House Kitchen (Ammo Up)",
-"[MMI1] Powerline Island Sandbox (Hatbox Tag)",
-"[MMI1] Powerline Island Left (Memory Vault)",
-"[MMI1] Powerline Island Right (Max Lives)",
-# The Neighborhood 2 [MMI2] 
-"[MMI2] Behind Book Depository (Steamer Trunk)",
-"[MMI2] Milkman Complete",
-
-# GLORIA'S THEATER
-# The Stage [THMS] 
-"[THMS] Near the Critic (Purse)",
-"[THMS] In the Audience (Ammo Up)",
-"[THMS] Below the Spotlight (Steamertrunk Tag)",
-"[THMS] Behind Stage (Purse Tag)",
-"[THMS] Behind Stage Cobweb (Suitcase)",
-"[THMS] Storage Room Floor (Vault)",
-"[THMS] Storage Room Left (Steamertrunk)",
-"[THMS] Storage Room Right Lower (Suitcase Tag)",
-"[THMS] Storage Room Right Upper (Candle # 1)",
-"[THMS] Bonita's Room",
-# The Catwalks [THCW] 
-"[THCW] Doghouse Slicers (Dufflebag Tag)",
-"[THCW] Big Platform 1 (Hatbox)",
-"[THCW] Big Platform 2 (Vault)",
-"[THCW] Big Platform 3 (One Up)",
-"[THCW] Big Platform Above (Hatbox Tag)",
-"[THCW] Next to Oatmeal (Dufflebag)",
-"[THCW] Candle Basket (Candle # 2)",
-"[THCW] Curtain Slide (Confusion Ammo Up)",
-"[THCW] Gloria's Theater Complete",
-
-# WATERLOO WORLD [WWMA] 
-"[WWMA] Fred's Room (Hatbox Tag)",
-"[WWMA] The Fireplace (Priceless Coin)",
-"[WWMA] Game Board (Suitcase Tag)",
-"[WWMA] Carpenter's Roof (Vault)",
-"[WWMA] Tightrope Room (Dufflebag)",
-"[WWMA] Outside Villager 1 House (One Up)",
-"[WWMA] Small Arch Top (Max Lives)",
-"[WWMA] Small Arch Below (Purse Tag)",
-"[WWMA] Top of Villager 2's House (Dufflebag Tag)",
-"[WWMA] Top of Villager 3's House (Ammo Up)",
-"[WWMA] Top of Knight's House (Confusion Ammo Up)",
-"[WWMA] Castle Tower (One Up)",
-"[WWMA] Castle Inside (Vault)",
-"[WWMA] Castle Wall (Steamertrunk)",
-"[WWMA] Under the Guillotine (Suitcase)",
-"[WWMA] Fred's House Basement (Hatbox)",
-"[WWMA] Blacksmith's Left Building (Purse)",
-"[WWMA] Blacksmith's Right Building (Steamertrunk Tag)",
-"[WWMA] Blacksmith's Haybale (The Musket)",
-"[WWMA] Help the Carpenter",
-"[WWMA] Help Villager 1",
-"[WWMA] Help the Knight",
-"[WWMA] Help Villager 2",
-"[WWMA] Help Villager 3",
-"[WWMA] Waterloo World Complete",
-
-# BLACK VELVETOPIA
-# Running Against the Bull [BVRB] 
-"[BVRB] Club Street Lady (Steamertrunk)",
-"[BVRB] Club Street Metal Balcony (Dufflebag Tag)",
-"[BVRB] Heart Street HIGH Balcony (Ammo Up)",
-"[BVRB] Alleyways Ledge (Hatbox Tag)",
-"[BVRB] Sewers Main (Vault)",
-"[BVRB] Club Street Gated (Steamer Trunk Tag)",
-"[BVRB] Burn the Logs (Dufflebag)",
-"[BVRB] The Garden (Vault)",
-"[BVRB] Near Diego's House (Max Lives)",
-"[BVRB] Diego's Bed (Suitcase Tag)",
-"[BVRB] Diego's Room (Hatbox)",
-"[BVRB] Diego's House Grindrail (Suitcase)",
-"[BVRB] Grindrail Balcony (Confusion Ammo Up)",
-
-# Edgar's Sanctuary [BVES] 
-"[BVES] Sanctuary Balcony (Purse Tag)",
-"[BVES] Sanctuary Ground (Purse)",
-"[BVES] Tiger Wrestler",
-"[BVES] Dragon Wrestler",
-"[BVES] Eagle Wrestler",
-"[BVES] Cobra Wrestler",
-"[BVES] Black Velvetopia Complete",
-
-# MEAT CIRCUS
-# Meat Circus Main [MCTC] 
-"[MCTC] Entrance Awning (Steamertrunk Tag)",
-"[MCTC] Crumbling Path (Steamertrunk)",
-"[MCTC] Crumbling Path End Right (Hatbox Tag)",
-"[MCTC] Crumbling Path End Left (Confusion Ammo Up)",
-"[MCTC] Ollie Escort Floor (Suitcase Tag)",
-"[MCTC] Ollie Escort Middle (Hatbox)",
-"[MCTC] Ollie Escort Top Left (Vault)",
-"[MCTC] Ollie Escort Top Right (Purse Tag)",
-"[MCTC] Tunnel of Love Start (Purse)",
-"[MCTC] Tunnel of Love Corner (Suitcase)",
-"[MCTC] Tunnel of Love Rail (Dufflebag Tag)",
-"[MCTC] Next to the Fat Lady (Dufflebag)",
-
-# DUMMY LOCATIONS
-# NOT COLLECTIBLE
-"DUMMY LOCATION 1 (NOT COLLECTIBLE)",
-"DUMMY LOCATION 2 (NOT COLLECTIBLE)",
-"DUMMY LOCATION 3 (NOT COLLECTIBLE)",
-
-]
