--- SeedGenerator/settings.ini
@@ -1,58 +1,0 @@
-[SeedSettings]
-startcobwebduster = True
-
-startlevitation = True
-
-startbutton = True
-
-randomizecobwebduster = True
-
-everylocationpossible = True
-
-instantdeath = False
-
-harderbutton = True
-
-[VictoryConditions]
-beatoleander = True
-
-rank101 = False
-
-brainhunt = True
-
-brainsrequired = 13
-
-scavengerhunt = False
-
-beatalllevels = False
-
-[QualityOfLife]
-fasterLO = True
-
-easymillarace = False
-
-earlyelevator = False
-
-mentalmagnet = True
-
-removetutorials = True
-
-easyflight = False
-
-[AdditionalFiles]
-spoilerlog = True
-
-createhints = True
-
-detailedhints = False
-
-#Minimum Value 0, Maximum Value 7
-majorhints = 3
-
-#Minimum Value 0, Maximum Value 8
-powerhints = 3
-
-#Minimum Value 0, Maximum Value 11
-minorhints = 3
-
-
