--- SeedGenerator/settingsinfo.txt
@@ -1,93 +1,0 @@
-
-For PsychoRando v1.0.0
-
-Generating a Seed:
-	Change settings in settings.ini to 'True' or 'False' as wanted, then run SeedGenerator.exe
-	Place RandoSeed.lua into the PsychoRando Scripts folder, overwrite old file.
-
-Using Presets:
-	Open the desired preset and copy the contents of the file (ctrl+a, ctrl+c), open settings.ini and paste over the old settings (ctrl+a, ctrl+v)
-
-[SeedSettings]
-
-	- _startcobwebduster_: If 'True', Adds the Cobweb Duster to your inventory at the start of the game,
-	opening access to more locations from the beginning. Has priority over 'randomizecobwebduster'
-
-	- _startlevitation_: If 'True', Adds one Levitation to starting inventory. Location in seed is replaced with
-	a Psicard.
-
-	- _startbutton_: If 'True', Adds Sasha's Button to starting inventory, unlocking Campgrounds. Location in seed is replaced with
-	a Psicard.
-
-	- _randomizecobwebduster_: If 'True', Removes Cobweb Duster from the shop and adds it to the item pool,
-	replacing a Psicard. Overridden by 'startcobwebduster'.
-
-	- _everylocationpossible_: Want to guarantee every location can be reached for a 100% completion run? 
-	Turn this on to make sure the generator knows no location can be left behind! Does not change player requirements for completing the game.
-
-	- _instantdeath_: WARNING!!! USE AT OWN RISK!!! If 'True', Raz will die instantly upon taking any amount of damage. 
-	Dream Fluffs and defensive measures are your best friend. Just don't make any mistakes.
-
-	- _harderbutton_: Want to make Sasha's Button harder to find? Use this setting to force its placement behind another progression item.
-
-
-[VictoryConditions]
-
-	- _beatoleander_: The original win condition, if 'True' this requires getting to the top of the Asylum
- 	and fighting the Coach Oleander Brain Tank before finishing the game.
-
-	- _rank101_: Race to max rank! If 'True', reaching Rank101 is mandatory to complete the game.
-	Time to crack some vaults, sort some baggage, and snag some figments!.
-
-	- _brainhunt_: Save the children! If 'True', player must find and break all 19 Brain Jars and 
-	return them all to Ford's Sanctuary for re-braining to beat the game. No child left behind.
-
-	- _scavengerhunt_: Ford DEMANDS you finish his Scavenger Hunt before you save the world. If 'True', 
-	all 16 scavenger hunt items must be gathered and redeemed with Ford in the GPC before finishing the game.
-
-	- _beatalllevels_: If 'True', you MUST complete every mental level in the game to access the final boss.
-	Does not include finishing the Asylum and Oleander Brain Tank.
-
-[QualityOfLife]
-
-	- _fasterLO_: Feeling sluggish in Lungfishopolis? If 'True', all of Raz's animations in Lungfishopolis are faster than normal, 
-	AND you can use Levitation to ride the levitation ball around! No floating or flying though.
-
-	- _easymillarace_: If losing to Bobby and having to restart Milla's race annoys you, set this to 'True'.
-	This setting removes Bobby from the race entirely, as well as increasing the game speed to 1.5x.
-
-	- _earlyelevator_: If 'True', the elevator at the Asylum entrance leading to Doctor Lobato's lab will be open 
-	if your inventory contains all three disguise pieces. This lets you skip the Upper Asylum floors if you've
-	already got all required powers and the cake.
-
-	- _mentalmagnet_: If 'True', start with the Mental Magnet badge in your inventory for easy health and ammo pickups.
-
-	- _removetutorials_: If 'True', removes ALL tutorial popups from the game (except the Ford Scavenger Hunt reminders).
-
-	- _easyflight_: Makes the levitation glitch obsolete, by letting Raz float UPWARDS instead of Downwards
-	while using psifloat. No more cramping hands or requiring high frame rates.
-
-[AdditionalFiles]
-
-	- _spoilerlog_: This setting is recommended to be left as 'True', generates SpoilerLog.txt, 
-	a list of all randomized items and their locations. Only turn this off to guarantee no cheating to find missing items.
-
-	- _createhints_: If 'True', generates Hints.txt, which lists a number of item locations meant to be viewed before
-	you start playing. The total number of hints and detail level of the hints is determined by the settings below.
-
-		- _detailedhints_: If 'True', Each hinted item will have a full location description matching the Spoiler Log.
-		If 'False', hinted items will only list the level they can be found in.
-		
-		- _majorhints_: Progression items required to access the Campgrounds, Asylum, and to beat the game.
-		Does not include Cobweb Duster or Sasha's Button if in starting inventory.
-		Minimum Value 0, Maximum Value 7
-
-		- _powerhints_: Raz's Psi Powers that help level progression and completing the game.
-		Does not include Levitation if in starting inventory. Only one of each available power can be hinted.
-		Minimum Value 0, Maximum Value 8
-
-		- _minorhints_: Progression items that are level specific, such as the Stop Sign or Megaphone, 
-		but not always required to complete the game.
-		Minimum Value 0, Maximum Value 11
-
-
